requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
aiohttp==3.11.18
python-multipart==0.0.20
PyMuPDF==1.26.4
python-dotenv==1.0.1
msrest==0.7.1
azure-ai-documentintelligence==1.0.2
python-telegram-bot==22.0
boto3==1.37.33
anthropic==0.49.0
openai==1.75.0
google-genai==1.12.1
